refactor(backend): move ClaudeArgsBuilder to modules/claude (phase 3 step 2)
